Initial implementation of package signature verification in WordPress.
This is written with some assumptions of the future work on WordPress.org and may change.
It also expects that PHP 5.4+ with ext/Sodium is in use, but won't fatal if not.

Signature Verification failures are reported back to WordPress.org for debugging purposes.

See https://core.trac.wordpress.org/ticket/39309